+++ Trabalho/TrabalhoFinal.cpp
@@ -2,38 +2,9 @@
 #include <string>
 #include <list>
 #include "Biblioteca.cpp"
-
-
+#include "menu.cpp"
 
 using namespace std;
-
-int opcao;
-int opcao_livro;
-
-// Menu principal
-int menu() {
-   system("cls");//limpa o cmd toda vez que volta para esse menu
-   cout <<"\n[ 1 ] Cadastros de livros \n";
-   cout <<"[ 2 ] Cadastro de usuarios \n";
-   cout <<"[ 3 ] Emprestimo e devolucao \n";
-   cout <<"[ 4 ] Busca e filtragem \n";
-   cout <<"[ 5 ] Relatorio \n";
-   cout <<"[ 6 ] Fim do App \n";
-   cout << "\nOpcao: ";
-   cin >> opcao;
-   return opcao; 
-}
-
-// Menu de cadastro de livros
-int menu_cadastro_livro() {
-   cout << "\n.:: MENU DA BIBLIOTECA ::.\n";
-   cout <<"[ 1 ] Adicionar Livro \n";
-   cout <<"[ 2 ] Exibir Livros \n";
-   cout <<"[ 3 ] Voltar ao menu principal \n";
-   cout << "\nOpcao: ";
-   cin >> opcao_livro;
-   return opcao_livro;
-}
 
 int main() {
     cout << ".:: SISTEMA BIBLIOTECA ::.";
@@ -57,6 +28,22 @@ int main() {
                     }
                 }
                 break;
+            case 2:
+                while (true){
+                    opcao_usuario = menu_cadastro_usuario();
+
+                    if(opcao_usuario == 1){
+                        cadastrarUsuario();//adicionar usuario
+                    }else if (opcao_usuario == 2) {
+                        break; // Volta ao menu principal
+                    }else {
+                        cout << "Opcao invalida!" << endl;
+                    }
+
+                }
+                
+                break;
+
 
             case 6: // Finalizar o aplicativo
                 cout << "Encerrando o aplicativo..." << endl;
